pyproject.toml
[project]
name = "pub-sub-trading-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python>=2.1.5",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
]
